% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{compute}
\alias{compute}
\title{Title}
\usage{
compute(a, b)
}
\arguments{
\item{a}{first operand}

\item{b}{second operand}
}
\value{
The product of "a" and "b"
}
\description{
Title
}
\examples{
compute(2, 4)
}
