% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World with Fruits}
\usage{
hello(x)
}
\arguments{
\item{x}{(string) name to be used.}
}
\value{
nothing, just printing a sentence.
}
\description{
This takes a name and then asks whether they like
a random fruit.
}
\examples{
hello("Ben")
}
