% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello2}
\alias{hello2}
\title{Alternative Hello Wolld}
\usage{
hello2(x)
}
\description{
Alternative Hello Wolld
}
\examples{
if (requireNamespace("dplyr")) {
  b <- data.frame(a = 1) |> dplyr::select(a)
  hello2(b$a)
}
}
