% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Example Data with Blup}
\format{
An object of class \code{data.frame} with 1 rows and 2 columns.
}
\usage{
mydata
}
\description{
Just contains 1 row.
}
\keyword{datasets}
