% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{compute}
\alias{compute}
\title{Multiplication}
\usage{
compute(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{second number.}
}
\value{
The product of `a` and `b`.
}
\description{
Multiplication
}
\examples{
compute(2, 4)
}
