% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello2}
\alias{hello2}
\title{Alternative Hello World}
\usage{
hello2(x)
}
\arguments{
\item{x}{(string) following Hello in the result.}
}
\value{
A string with "Hello x Daniel"
}
\description{
Alternative Hello World
}
\examples{
if (requireNamespace("dplyr")) {
  b <- data.frame(a = 1) |> dplyr::select(a)
  hello2(b$a)
}

}
